+++ verification/src/lib.rs
@@ -25,7 +25,7 @@ pub fn verify(vf: &mut VerifierContext, to_verify: &ClassBackedView, loader: Loa
 
 #[derive(Debug)]
 pub struct StackMap {
-    pub offset: usize,
+    pub offset: u16,
     pub map_frame: Frame,
 }
 
@@ -34,7 +34,7 @@ pub struct VerifierContext<'l> {
     pub classfile_getter: Arc<dyn ClassFileGetter + 'l>,
     // pub classes: &'l ,
     pub current_loader: LoaderName,
-    pub verification_types: HashMap<u16, HashMap<usize, Frame>>,
+    pub verification_types: HashMap<u16, HashMap<u16, Frame>>,
     pub debug: bool,
 }
